% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_ingest.R
\name{cloud_ingest}
\alias{cloud_ingest}
\title{cloud_ingest}
\usage{
cloud_ingest(
  path_to_ras_dbase,
  top_of_dir_to_scrape,
  code_to_place_in_source,
  proj_override = NULL,
  vdat_trans = FALSE,
  quiet = FALSE,
  verbose = FALSE,
  quick_check = FALSE,
  quick_hull = FALSE,
  overwrite = FALSE,
  free_treads = 4,
  refresh = TRUE
)
}
\arguments{
\item{path_to_ras_dbase}{PARAM_DESCRIPTION}

\item{top_of_dir_to_scrape}{PARAM_DESCRIPTION}

\item{code_to_place_in_source}{PARAM_DESCRIPTION}

\item{proj_override}{PARAM_DESCRIPTION, Default: NULL}

\item{vdat_trans}{PARAM_DESCRIPTION, Default: FALSE}

\item{quiet}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{quick_check}{PARAM_DESCRIPTION, Default: FALSE}

\item{quick_hull}{PARAM_DESCRIPTION, Default: FALSE}

\item{overwrite}{PARAM_DESCRIPTION, Default: FALSE}

\item{free_treads}{PARAM_DESCRIPTION, Default: 4}

\item{refresh}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_sub}}, \code{\link[stringr]{str_detect}}
 \code{\link[aws.s3]{get_bucket}}, \code{\link[aws.s3]{delete_object}}, \code{\link[aws.s3]{put_object}}
 \code{\link[glue]{glue}}
 \code{\link[utils]{glob2rx}}
 \code{\link[sf]{st_crs}}, \code{\link[sf]{st_coordinates}}, \code{\link[sf]{st_as_sf}}, \code{\link[sf]{st_transform}}, \code{\link[sf]{st_write}}
 \code{\link[dplyr]{select}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{distinct}}
 \code{\link[data.table]{data.table-package}}, \code{\link[data.table]{fwrite}}
 \code{\link[sfheaders]{sf_linestring}}
 \code{\link[lwgeom]{st_startpoint}}
 \code{\link[holyhull]{holyhull}}
 \code{\link[nhdplusTools]{get_nhdplus}}
 \code{\link[AOI]{aoi_get}}
 \code{\link[arrow]{write_parquet}}
 \code{\link[parallel]{detectCores}}, \code{\link[parallel]{makeCluster}}
 \code{\link[doParallel]{registerDoParallel}}
 \code{\link[foreach]{foreach}}
}
