% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedupe.R
\name{dedupe}
\alias{dedupe}
\title{dedupe}
\usage{
dedupe()
}
\value{
updated ras catalog
}
\description{
de-duplicate otherwise duplicate ras catalog entries
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 #RRASSLER::dedupe()
 }
}
}
