% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datum_from_crs.R
\name{get_datum_from_crs}
\alias{get_datum_from_crs}
\title{get_datum_from_crs}
\usage{
get_datum_from_crs(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
list of CRS and vertical unit
}
\description{
attempts to parse datum and datum unit from file based on WKT
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 get_datum_from_crs("EPSG:6349")

 }
}
}
\seealso{
[st_crs][sf::st_crs]
}
