% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_manual_record.R
\name{insert_manual_record}
\alias{insert_manual_record}
\title{insert_manual_record}
\usage{
insert_manual_record(
  in_file = NULL,
  path_to_ras_dbase = NULL,
  xs_shapefile,
  river_shapefile,
  geometric_realization_override = "g01",
  code_to_place_in_source = NULL,
  proj_override = NULL,
  is_quiet = FALSE,
  is_verbose = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{in_file}{the path to the file on disk that we want to ingest, Default: NULL}

\item{path_to_ras_dbase}{A path to a directory to write your RRASSLED catalog to, Default: NULL}

\item{xs_shapefile}{the path to the exported cross section shapefile}

\item{river_shapefile}{the path to the exported river shapefile}

\item{geometric_realization_override}{the geometric realization you are parsing out as a sting with leading g character ie g01, g02..., Default: 'g01'}

\item{code_to_place_in_source}{a code to place in the metadata as the owner of the model, Default: NULL}

\item{proj_override}{a string to override projection information should none be found, Default: NULL}

\item{is_quiet}{flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE}

\item{is_verbose}{flag to determine whether internal print statements (i.e. cross section parsing, vdat trans, file info) are suppressed, TRUE to show these messages and FALSE to suppress them, Default: FALSE}

\item{overwrite}{overwrite overwrite files if we find identical models, Default: FALSE}
}
\value{
a set of files in a newly RRASSLE'd record.
}
\description{
forces a manually exported version of a model into the RRASSLED structure
}
\details{
To export open the 'RAS Mapper' (GIS Tools > RAS Mapper) and right click on the 'Rivers' and 'Cross sections' layers and 'Export Layer > Save Layer to Shapefile' to a shapefile (I append mine with the model name). See https://github.com/NOAA-OWP/RRASSLER/blob/dev/man/figures/_export_for_insert_manual.png for a verbose example.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 ras_dbase <- "./inst/extdata/sample_output/ras_catalog/"
 host_path = file.path(".../ras_catalog/_temp/supercustommodel/", fsep = .Platform$file.sep)

 insert_manual_record(in_file = file.path(host_path, "Ohio2018a.prj",fsep = .Platform$file.sep),path_to_ras_dbase = ras_dbase,xs_shapefile = file.path(host_path, "Ohio2018a.RASexport_xs_z.shp", fsep = .Platform$file.sep),river_shapefile = file.path(host_path, "Ohio2018a.RASexport_stream.shp", fsep = .Platform$file.sep),geometric_realization_override = "g02",code_to_place_in_source = "RFC",proj_override = file.path(host_path, "Ohio2018a.RASexport_stream.shp", fsep = .Platform$file.sep),is_quiet = FALSE,is_verbose = TRUE,overwrite = FALSE)
 }
}
}
\seealso{
\code{\link[utils]{glob2rx}}, \code{\link[utils]{read.table}}
 \code{\link[glue]{glue}}
 \code{\link[sf]{st_read}}, \code{\link[sf]{st_crs}}, \code{\link[sf]{st_transform}}, \code{\link[sf]{st_cast}}, \code{\link[sf]{st_coordinates}}, \code{\link[sf]{geos_measures}}, \code{\link[sf]{st_geometry}}, \code{\link[sf]{st_write}}
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{row_number}}, \code{\link[dplyr]{c("rowwise", "rowwise")}}, \code{\link[dplyr]{group_by}}
 \code{\link[lwgeom]{st_startpoint}}
 \code{\link[sfheaders]{sf_polygon}}, \code{\link[sfheaders]{sf_linestring}}
 \code{\link[units]{units}}
 \code{\link[stringr]{str_detect}}
 \code{\link[data.table]{data.table-package}}, \code{\link[data.table]{fwrite}}
 \code{\link[arrow]{write_parquet}}

Other ingest: 
\code{\link{cloud_ingest_record}()},
\code{\link{disk_ingest_record}()},
\code{\link{ingest_FEMA6_BLE}()},
\code{\link{ingest_into_database}()}
}
\concept{ingest}
