% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_dbase.R
\name{move_dbase}
\alias{move_dbase}
\title{move_dbase}
\usage{
move_dbase(from = NULL, to = NULL)
}
\value{
a copy of the database in the new location
}
\description{
Bulk database mover
}
\details{
dev
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 move_dbase(from = NULL, to = NULL)
 }
}
}
\seealso{
Other helper: 
\code{\link{marco}()},
\code{\link{print_error_block}()},
\code{\link{print_warning_block}()},
\code{\link{url_exists}()},
\code{\link{util_ras2plucker}()}
}
\concept{helper}
