% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_and_plot_old.R
\name{parse_and_plot_old}
\alias{parse_and_plot_old}
\title{FUNCTION_TITLE}
\usage{
parse_and_plot_old(
  gfile_path = NULL,
  path_to_ras_dbase = NULL,
  excel_row = NULL,
  units_overwrite = NULL,
  proj_overwrite = NULL,
  vdat_trans = FALSE,
  plots_wanted = "g",
  outpath = NULL,
  overwrite = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{gfile_path}{PARAM_DESCRIPTION, Default: NULL}

\item{path_to_ras_dbase}{PARAM_DESCRIPTION, Default: NULL}

\item{excel_row}{PARAM_DESCRIPTION, Default: NULL}

\item{units_overwrite}{PARAM_DESCRIPTION, Default: NULL}

\item{proj_overwrite}{PARAM_DESCRIPTION, Default: NULL}

\item{vdat_trans}{PARAM_DESCRIPTION, Default: FALSE}

\item{plots_wanted}{PARAM_DESCRIPTION, Default: 'g'}

\item{outpath}{PARAM_DESCRIPTION, Default: NULL}

\item{overwrite}{PARAM_DESCRIPTION, Default: TRUE}

\item{quiet}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
 \code{\link[utils]{glob2rx}}
 \code{\link[sf]{st_crs}}
 \code{\link[sfheaders]{sf_linestring}}
 \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{geom_path}}, \code{\link[ggplot2]{geom_point}}, \code{\link[ggplot2]{ggtheme}}, \code{\link[ggplot2]{labs}}, \code{\link[ggplot2]{CoordSf}}
 \code{\link[cowplot]{plot_grid}}, \code{\link[cowplot]{ggdraw}}, \code{\link[cowplot]{draw_label}}, \code{\link[cowplot]{save_plot}}
}
