% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_warning_block.R
\name{print_warning_block}
\alias{print_warning_block}
\title{print_warning_block}
\usage{
print_warning_block()
}
\value{
print output
}
\description{
A warning block helper
}
\details{
line art to draw eyes to printing notices.  Used when I want to alert to an uncritical error or trace RRASSLER behavior.
}
\examples{
RRASSLER::print_warning_block()
}
\seealso{
Other helper: 
\code{\link{load_catalog_csv_as_DT}()},
\code{\link{marco}()},
\code{\link{print_error_block}()},
\code{\link{process_ras_hdf_to_rasxyz}()},
\code{\link{url_exists}()},
\code{\link{util_unzip}()}
}
\concept{helper}
