% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ras_hdf_to_rasxyz.R
\name{process_ras_hdf_to_rasxyz}
\alias{process_ras_hdf_to_rasxyz}
\title{process_ras_hdf_to_rasxyz}
\usage{
process_ras_hdf_to_rasxyz(
  geom_path,
  units,
  proj_string,
  in_epoch_override = as.integer(as.POSIXct(Sys.time())),
  out_epoch_override = as.integer(as.POSIXct(Sys.time())),
  vdat = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{geom_path}{path to a file to parse}

\item{units}{units found in the project, "English Units" or "SI Units"}

\item{proj_string}{a projection string to apply}

\item{in_epoch_override}{vdatum parameter input epoch, Default: as.integer(as.POSIXct(Sys.time()))}

\item{out_epoch_override}{vdatum parameter output epoch, Default: as.integer(as.POSIXct(Sys.time()))}

\item{vdat}{a flag to dictate whether or not to apply a vdatum transformation, TRUE to apply, FALSE to skip, Default: FALSE}

\item{quiet}{flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE}
}
\value{
a point database and notes about processing
}
\description{
process a ras g##.hdf file into a stream oriented and classified xyz format
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 # ghdf_path <- "./inst/extdata/sample_ras/FEMA-R6-BLE-sample-dataset/12090301/12090301_models/Model/Alum Creek-Colorado River/ALUM 006/ALUM 006.g01.hdf"
 ghdf_path <- fs::path_package("extdata/shapes.fgb", package = "mypkg")
 pts <- process_ras_hdf_to_rasxyz(geom_path = ghdf_path,units = "English Units",proj_string = "EPSG:2277",vdat = FALSE,quiet = FALSE)
 }
}
}
\seealso{
[str_detect][stringr::str_detect]
 [stri_sub][stringi::stri_sub]
 [h5read][rhdf5::h5read]
 [sf_linestring][sfheaders::sf_linestring]
 [st_sf][sf::st_sf], [st_cast][sf::st_cast], [st_crs][sf::st_crs], [st_transform][sf::st_transform], [st_coordinates][sf::st_coordinates], [st_as_sf][sf::st_as_sf], [st_set_crs][sf::st_set_crs], [st_length][sf::st_length]
 [decimal_date][lubridate::decimal_date], [ymd][lubridate::ymd]
 [GET][httr::GET], [http_error][httr::http_error], [content][httr::content]
 [fill][tidyr::fill]
 [st_linesubstring][lwgeom::st_linesubstring], [st_endpoint][lwgeom::st_endpoint]
 [glue][glue::glue]

Other helper: 
\code{\link{load_catalog_csv_as_DT}()},
\code{\link{marco}()},
\code{\link{print_error_block}()},
\code{\link{print_warning_block}()},
\code{\link{url_exists}()},
\code{\link{util_unzip}()}
}
\concept{helper}
