% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprocess_table_row.R
\name{reprocess_table_row}
\alias{reprocess_table_row}
\title{reprocess_table_row}
\usage{
reprocess_table_row(path_to_ras_dbase, excel_row)
}
\arguments{
\item{path_to_ras_dbase}{The path to the top level folder of the ras_catalog structure}

\item{excel_row}{if you opened this up in excel (1 based index) the row you want to process}
}
\value{
updates a row to hopefully place the model in a better space
}
\description{
regenerate a row in the catalog
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[glue][glue::glue]
 [fwrite][data.table::fwrite]
 [sf_linestring][sfheaders::sf_linestring]
 [st_set_crs][sf::st_set_crs], [st_crs][sf::st_crs], [st_coordinates][sf::st_coordinates], [st_as_sf][sf::st_as_sf], [st_transform][sf::st_transform], [st_convex_hull][sf::st_convex_hull], [st_union][sf::st_union], [st_write][sf::st_write]
 [st_endpoint][lwgeom::st_endpoint], [st_startpoint][lwgeom::st_startpoint]
 [group_by][dplyr::group_by], [distinct][dplyr::distinct]
 [get_nhdplus][nhdplusTools::get_nhdplus]
 [aoi_get][AOI::aoi_get]
 [write_parquet][arrow::write_parquet]
}
