% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{scraper}
\alias{scraper}
\title{scraper}
\usage{
scraper(path_to_ras_dbase = TRUE)
}
\arguments{
\item{path_to_ras_dbase}{The path to the top level folder of the ras_catalog structure, Default: TRUE}
}
\value{
a new set of model footprints to map
}
\description{
silly handle around unscraped databases
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{as.data.table}}, \code{\link[data.table]{data.table-package}}, \code{\link[data.table]{rbindlist}}, \code{\link[data.table]{fwrite}}
 \code{\link[arrow]{write_parquet}}
 \code{\link[sf]{st_write}}
}
