% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotgun_proj_test.R
\name{shotgun_proj_test}
\alias{shotgun_proj_test}
\title{shotgun_proj_test}
\usage{
shotgun_proj_test(
  path_to_model_g,
  out_path,
  proj_test_list = c("EPSG:2277", "ESRI:102739"),
  units = "Foot",
  in_epoch_override = as.integer(as.POSIXct(Sys.time())),
  out_epoch_override = as.integer(as.POSIXct(Sys.time())),
  vdat_trans = FALSE,
  quiet = FALSE,
  default_g = FALSE,
  try_both = TRUE
)
}
\arguments{
\item{path_to_model_g}{PARAM_DESCRIPTION}

\item{out_path}{PARAM_DESCRIPTION}

\item{proj_test_list}{PARAM_DESCRIPTION, Default: c("EPSG:2277", "ESRI:102739")}

\item{units}{PARAM_DESCRIPTION, Default: 'Foot'}

\item{in_epoch_override}{PARAM_DESCRIPTION, Default: as.integer(as.POSIXct(Sys.time()))}

\item{out_epoch_override}{PARAM_DESCRIPTION, Default: as.integer(as.POSIXct(Sys.time()))}

\item{vdat_trans}{PARAM_DESCRIPTION, Default: FALSE}

\item{quiet}{PARAM_DESCRIPTION, Default: FALSE}

\item{default_g}{PARAM_DESCRIPTION, Default: FALSE}

\item{try_both}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 RRASSLER::shotgun_proj_test(path_to_model_g="G:/data/ras_catalog/_temp/BLE/12090301/12090301_models/Model/Walnut Creek-Colorado River/CEDAR CREEK/CEDAR CREEK.g01",out_path="G:/data/ras_catalog/_temp/test",units = "Foot",proj_test_list = c("EPSG:2277","ESRI:102739"),in_epoch_override = as.integer(as.POSIXct(Sys.time())),out_epoch_override = as.integer(as.POSIXct(Sys.time())),vdat_trans = FALSE,quiet = FALSE,default_g = FALSE,try_both = TRUE)
 }
}
}
\seealso{
\code{\link[glue]{glue}}
 \code{\link[sfheaders]{sf_linestring}}
 \code{\link[sf]{st_crs}}, \code{\link[sf]{st_coordinates}}, \code{\link[sf]{st_as_sf}}, \code{\link[sf]{st_transform}}, \code{\link[sf]{st_write}}
 \code{\link[lwgeom]{st_startpoint}}
 \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{distinct}}
 \code{\link[holyhull]{holyhull}}
}
