% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_g_to_geom_pts.R
\name{util_g_to_geom_pts}
\alias{util_g_to_geom_pts}
\title{util_g_to_geom_pts}
\usage{
util_g_to_geom_pts(geom_path, units, proj_string, quiet = FALSE)
}
\arguments{
\item{geom_path}{PARAM_DESCRIPTION}

\item{units}{PARAM_DESCRIPTION}

\item{proj_string}{PARAM_DESCRIPTION}

\item{quiet}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[utils]{glob2rx}}, \code{\link[utils]{read.table}}
 \code{\link[glue]{glue}}
 \code{\link[data.table]{as.data.table}}
 \code{\link[stringr]{str_split}}, \code{\link[stringr]{str_trim}}, \code{\link[stringr]{str_flatten}}
 \code{\link[sfheaders]{sf_linestring}}
 \code{\link[sf]{sf}}, \code{\link[sf]{st_cast}}
 \code{\link[tidyr]{fill}}
}
