% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_units_from_g.R
\name{util_units_from_g}
\alias{util_units_from_g}
\title{util_units_from_g}
\usage{
util_units_from_g(gpath, is_quiet = FALSE)
}
\arguments{
\item{gpath}{path to geometry file}

\item{is_quiet}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
units for RRASSLER processing
}
\description{
helper to grab units from geom path
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[utils]{glob2rx}}
 \code{\link[glue]{glue}}
}
