% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ras2fim_helper.R, R/util_unzip.R
\name{util_ras2plucker}
\alias{util_ras2plucker}
\alias{util_unzip}
\title{util_ras2plucker}
\usage{
util_ras2plucker(
  path_to_ras_dbase = NULL,
  AOI = NULL,
  out_name = NULL,
  overwrite = FALSE,
  is_verbose = TRUE,
  is_quiet = FALSE
)

util_unzip(zippath, is_quiet = FALSE)
}
\arguments{
\item{is_quiet}{flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE}

\item{zippath}{path to zip}
}
\value{
unzipped dir

unzipped dir
}
\description{
pulls models from model catalog in a similar fashion to https://github.com/NOAA-OWP/ras2fim/blob/dev/tools/s3_get_models.py and compatable with ras2fim v2

unzips nested zip files
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 RRASSLER::util_unzip(file.path(database_path,"_temp","BLE",HUCID,glue::glue("{HUCID}_models.zip"),fsep = .Platform$file.sep),is_quiet = is_quiet)
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 RRASSLER::util_unzip(file.path(database_path,"_temp","BLE",HUCID,glue::glue("{HUCID}_models.zip"),fsep = .Platform$file.sep),is_quiet = is_quiet)
 }
}
}
\seealso{
\code{\link[stringr]{str_sub}}
 \code{\link[utils]{unzip}}
 \code{\link[glue]{glue}}
 \code{\link[dplyr]{setops}}

\code{\link[stringr]{str_sub}}
 \code{\link[utils]{unzip}}
 \code{\link[glue]{glue}}
 \code{\link[dplyr]{setops}}

Other helper: 
\code{\link{marco}()},
\code{\link{move_dbase}()},
\code{\link{print_error_block}()},
\code{\link{print_warning_block}()},
\code{\link{url_exists}()}

Other helper: 
\code{\link{marco}()},
\code{\link{move_dbase}()},
\code{\link{print_error_block}()},
\code{\link{print_warning_block}()},
\code{\link{url_exists}()}
}
\concept{helper}
