% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs_to_transect_picker.R
\name{xs_to_transect_picker}
\alias{xs_to_transect_picker}
\title{xs_to_transect_picker}
\usage{
xs_to_transect_picker(
  path_to_ras_dbase = NULL,
  line_select_subset = NULL,
  hf_lines,
  reference_divides,
  reference_flowlines,
  existing_transects,
  existing_points,
  overwrite = FALSE,
  is_verbose = TRUE,
  test_limit = NULL
)
}
\arguments{
\item{path_to_ras_dbase}{The path to the folder in which you are building your catalog, Default: NULL}

\item{line_select_subset}{PARAM_DESCRIPTION, Default: NULL}

\item{hf_lines}{Hydrofabric lines}

\item{reference_divides}{reference divide paths}

\item{reference_flowlines}{reference flowline paths}

\item{existing_transects}{reference transect paths}

\item{existing_points}{reference points paths}

\item{overwrite}{overwrite files if we find identical models, Default: FALSE}

\item{is_verbose}{flag to determine whether internal print statements (i.e. cross section parsing, vdat trans, file info) are suppressed, TRUE to show these messages and FALSE to suppress them, Default: FALSE}

\item{test_limit}{Number of transects to use, Default: NULL}
}
\value{
selected HEC-RAS lines in hd3d form
}
\description{
Picks transects
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 xs_to_transect_picker(path_to_ras_dbase = NULL,line_select_subset = NULL,hf_lines,overwrite = FALSE,is_verbose = TRUE)
 }
}
}
\seealso{
\code{\link[sf]{st_read}}, \code{\link[sf]{st_as_sf}}, \code{\link[sf]{st_crs}}, \code{\link[sf]{st_transform}}, \code{\link[sf]{sf}}, \code{\link[sf]{sfc}}, \code{\link[sf]{st_cast}}, \code{\link[sf]{st_line_sample}}, \code{\link[sf]{st_as_sfc}}, \code{\link[sf]{st_bbox}}, \code{\link[sf]{geos_binary_ops}}, \code{\link[sf]{st_nearest_feature}}, \code{\link[sf]{geos_measures}}, \code{\link[sf]{st_coordinates}}, \code{\link[sf]{st_write}}
 \code{\link[arrow]{read_parquet}}, \code{\link[arrow]{open_dataset}}, \code{\link[arrow]{write_parquet}}
 \code{\link[dplyr]{compute}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{row_number}}
 \code{\link[glue]{glue}}
}
