% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_catalog_fields.R
\name{append_catalog_fields}
\alias{append_catalog_fields}
\title{FUNCTION_TITLE}
\usage{
append_catalog_fields(
  path_to_ras_dbase = NULL,
  out_name = NULL,
  dat_path = NULL,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{path_to_ras_dbase}{PARAM_DESCRIPTION, Default: NULL}

\item{out_name}{PARAM_DESCRIPTION, Default: NULL}

\item{dat_path}{PARAM_DESCRIPTION, Default: NULL}

\item{overwrite}{PARAM_DESCRIPTION, Default: FALSE}

\item{quiet}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
 \code{\link[dplyr]{mutate}}
 \code{\link[sf]{s2}}, \code{\link[sf]{st_transform}}, \code{\link[sf]{st_read}}, \code{\link[sf]{st_crs}}
 \code{\link[data.table]{fwrite}}
}
