% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_catalog_fields.R
\name{append_catalog_fields}
\alias{append_catalog_fields}
\title{append_catalog_fields}
\usage{
append_catalog_fields(
  path_to_ras_dbase,
  out_name = "OWP_ras_catalog.csv",
  dat_path = "/home/rstudio/g/Dropbox/root/database/hosted/water/HUC8.fgb",
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{path_to_ras_dbase}{The path to the top level folder of the ras_catalog structure}

\item{out_name}{PARAM_DESCRIPTION, Default: 'OWP_ras_catalog.csv'}

\item{dat_path}{PARAM_DESCRIPTION, Default: '/home/rstudio/g/Dropbox/root/database/hosted/water/HUC8.fgb'}

\item{overwrite}{PARAM_DESCRIPTION, Default: FALSE}

\item{quiet}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Append calculated fields and helper values to new catalog
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 # RRASSLER::append_catalog_fields(path_to_ras_dbase="/home/rstudio/g/data/ras_dbase",out_name="OWP_ras_catalog.csv",overwrite=FALSE,quiet=FALSE)
 }
}
}
\seealso{
\code{\link[glue]{glue}}
 \code{\link[sf]{st_transform}}, \code{\link[sf]{st_read}}, \code{\link[sf]{st_crs}}
}
