% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_catalog_fields.R
\name{append_catalog_fields}
\alias{append_catalog_fields}
\title{append_catalog_fields}
\usage{
append_catalog_fields(
  path_to_ras_dbase = NULL,
  out_name = NULL,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{path_to_ras_dbase}{the path to the folder in which you are building your catalog, Default: NULL}

\item{out_name}{the name of the csv you want to generate, Default: NULL}

\item{overwrite}{flag to dictate whether or not to overwrite the out_name, should it exist. set to TRUE to delete and (re)generate, FALSE to safely exit, Default: FALSE}

\item{quiet}{flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE}
}
\value{
a new csv with helper columns
}
\description{
adds helper fields to accounting.csv
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
 \code{\link[dplyr]{mutate}}
 \code{\link[sf]{s2}}, \code{\link[sf]{st_transform}}, \code{\link[sf]{st_read}}, \code{\link[sf]{st_crs}}
 \code{\link[data.table]{fwrite}}
}
