% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_catalog_fields.R
\name{append_catalog_fields}
\alias{append_catalog_fields}
\title{append_catalog_fields}
\usage{
append_catalog_fields(
  path_to_ras_dbase = NULL,
  out_name = NULL,
  overwrite = FALSE,
  is_verbose = TRUE,
  HUC8_override = NULL
)
}
\arguments{
\item{path_to_ras_dbase}{the path to the folder in which you are building your catalog, Default: NULL}

\item{out_name}{the name of the csv you want to generate, Default: NULL}

\item{overwrite}{flag to dictate whether or not to overwrite the out_name, should it exist. set to TRUE to delete and (re)generate, FALSE to safely exit, Default: FALSE}

\item{is_verbose}{flag to determine whether print statements are suppressed, TRUE to show messages and FALSE to surpress them, Default: TRUE}

\item{HUC8_override}{a path to the spatial key if you need to run this over a temp dir for eg ras2fim, Default: NULL}
}
\value{
a new csv with helper columns
}
\description{
adds helper fields to accounting.csv.
}
\details{
TRUE
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE
 RRASSLER::append_catalog_fields(path_to_ras_dbase = "G:/data/ras_catalog",out_name = "OWP_ras_model_catalog.csv",overwrite = FALSE,is_verbose = TRUE)
 }
}
}
\seealso{
\code{\link[glue]{glue}}
 \code{\link[dplyr]{mutate}}
 \code{\link[sf]{s2}}, \code{\link[sf]{st_transform}}, \code{\link[sf]{st_read}}, \code{\link[sf]{st_crs}}
 \code{\link[data.table]{fwrite}}

Other post-process: 
\code{\link{refresh_master_files}()}
}
\concept{post-process}
