% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_ingest_record.R
\name{cloud_ingest_record}
\alias{cloud_ingest_record}
\title{cloud_ingest_record}
\usage{
cloud_ingest_record(
  in_file = NULL,
  ras_dbase = NULL,
  root_bucket = NULL,
  code_to_place_in_source = NULL,
  proj_override = NULL,
  vdat_trans = FALSE,
  quiet = FALSE,
  verbose = FALSE,
  quick_check = FALSE,
  quick_hull = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{in_file}{PARAM_DESCRIPTION, Default: NULL}

\item{ras_dbase}{PARAM_DESCRIPTION, Default: NULL}

\item{root_bucket}{PARAM_DESCRIPTION, Default: NULL}

\item{code_to_place_in_source}{PARAM_DESCRIPTION, Default: NULL}

\item{proj_override}{PARAM_DESCRIPTION, Default: NULL}

\item{vdat_trans}{PARAM_DESCRIPTION, Default: FALSE}

\item{quiet}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{quick_check}{PARAM_DESCRIPTION, Default: FALSE}

\item{quick_hull}{PARAM_DESCRIPTION, Default: FALSE}

\item{overwrite}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
