% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_hull_to_hydrofabric_value.R
\name{crosswalk_hull_to_hydrofabric_value}
\alias{crosswalk_hull_to_hydrofabric_value}
\title{crosswalk_hull_to_hydrofabric_value}
\usage{
crosswalk_hull_to_hydrofabric_value(hull, river)
}
\arguments{
\item{hull}{the model cross section hulls}

\item{river}{the river streamlines}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
crosswalk_hull_to_hydrofabric_value
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[nhdplusTools]{get_nhdplus}}
 \code{\link[AOI]{aoi_get}}
 \code{\link[sf]{st_transform}}, \code{\link[sf]{st_crs}}
}
\concept{helpers}
