% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedupe.R
\name{dedupe}
\alias{dedupe}
\title{dedupe}
\usage{
dedupe(path_to_ras_dbase = NULL, quiet = FALSE)
}
\arguments{
\item{path_to_ras_dbase}{the path to the folder in which you are building your catalog, Default: NULL}

\item{quiet}{flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE}
}
\value{
a list of files with otherwise identical geometries
}
\description{
de duplicates RAS models from the catalog
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
