% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disk_ingest_record.R
\name{disk_ingest_record}
\alias{disk_ingest_record}
\title{disk_ingest_record}
\usage{
disk_ingest_record(
  in_file = NULL,
  path_to_ras_dbase = NULL,
  code_to_place_in_source = NULL,
  proj_override = NULL,
  apply_vdat_trans = FALSE,
  is_quiet = FALSE,
  is_verbose = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{in_file}{the path to the file on disk that we want to ingest, Default: NULL}

\item{code_to_place_in_source}{a code to place in the metadata as the owner of the model, Default: NULL}

\item{proj_override}{a string to override projection information should none be found, Default: NULL}

\item{apply_vdat_trans}{Should VDATUM be applied to the HEC-RAS model geometry.  See https://vdatum.noaa.gov/, Default: FALSE}

\item{is_quiet}{flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE}

\item{is_verbose}{flag to determine whether internal print statements (i.e. cross section parsing, vdat trans, file info) are suppressed, TRUE to show these messages and FALSE to suppress them, Default: FALSE}

\item{overwrite}{overwrite files if we find identical models, Default: FALSE}

\item{ras_dbase}{A path to a directory to write your RRASSLED catalog to., Default: NULL}
}
\value{
a set of files in a newly RRASSLE'd record.
}
\description{
add a file as a record in the RRASSLED structure
}
\details{
the disk version of the ingest process
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 # ras_dbase <- "./inst/extdata/sample_output/ras_catalog/"
 # dir_to_scrape <- "./inst/extdata/sample_ras/FEMA-R6-BLE-sample-dataset/"
 dir_to_scrape <- fs::path_package("extdata/shapes.fgb", package = "mypkg")
 list_of_prj_files <- list.files(dir_to_scrape,pattern = glob2rx("*.prj$"),full.names = TRUE,ignore.case = TRUE,recursive = TRUE)
 disk_ingest_record(in_file = list_of_prj_files[1], path_to_ras_dbase = ras_dbase,code_to_place_in_source = "test",proj_override = "EPSG:2277",apply_vdat_trans = FALSE,is_quiet = FALSE,is_verbose = FALSE,overwrite = FALSE)

 dir_to_scrape <- "./inst/extdata/sample_ras/ras2fim-sample-dataset/"
 list_of_prj_files <- list.files(dir_to_scrape,pattern = glob2rx("*.prj$"),full.names = TRUE,ignore.case = TRUE,recursive = TRUE)
 disk_ingest_record(in_file = list_of_prj_files[1], path_to_ras_dbase = ras_dbase,code_to_place_in_source = "test",proj_override = "EPSG:26915",apply_vdat_trans = FALSE,is_quiet = FALSE,is_verbose = TRUE,overwrite = FALSE)
 }
}
}
\seealso{
\code{\link[utils]{glob2rx}}
 \code{\link[glue]{glue}}
 \code{\link[sf]{st_crs}}, \code{\link[sf]{st_coordinates}}, \code{\link[sf]{st_write}}
 \code{\link[stringr]{str_sub}}, \code{\link[stringr]{str_detect}}
 \code{\link[data.table]{data.table-package}}, \code{\link[data.table]{fwrite}}
 \code{\link[sfheaders]{sf_linestring}}, \code{\link[sfheaders]{sf_polygon}}
 \code{\link[lwgeom]{st_startpoint}}
 \code{\link[arrow]{write_parquet}}

Other ingest: 
\code{\link{cloud_ingest_record}()},
\code{\link{ingest_FEMA6_BLE}()},
\code{\link{ingest_into_database}()}
}
\concept{ingest}
