% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrankenFIM_scrape_and_unpack_ble.R
\name{FrankenFIM_scrape_and_unpack_ble}
\alias{FrankenFIM_scrape_and_unpack_ble}
\title{scrape_and_unpack_ble}
\usage{
FrankenFIM_scrape_and_unpack_ble(
  database_path = NULL,
  HUCID = NULL,
  quiet = TRUE,
  overwrite = FALSE,
  full = FALSE
)
}
\arguments{
\item{database_path}{PARAM_DESCRIPTION, Default: NULL}

\item{HUCID}{PARAM_DESCRIPTION, Default: NULL}

\item{quiet}{PARAM_DESCRIPTION, Default: TRUE}

\item{overwrite}{PARAM_DESCRIPTION, Default: FALSE}

\item{full}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
unzipped dir of BLE data
}
\description{
wrapper around BLE handles
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 RRASSLER::FrankenFIM_scrape_and_unpack_ble(database_path=NULL,HUCID=NULL,quiet=TRUE,overwrite=FALSE,full=FALSE)
 }
}
}
\seealso{
\code{\link[httr]{HEAD}}, \code{\link[httr]{GET}}, \code{\link[httr]{status_code}}, \code{\link[httr]{write_disk}}
 \code{\link[sf]{st_transform}}, \code{\link[sf]{st_read}}, \code{\link[sf]{st_crs}}
 \code{\link[glue]{glue}}
 \code{\link[stringr]{str_sub}}
 \code{\link[utils]{unzip}}
 \code{\link[dplyr]{setops}}
}
