% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_FEMA6_BLE.R
\name{ingest_FEMA6_BLE}
\alias{ingest_FEMA6_BLE}
\title{ingest_FEMA6_BLE}
\usage{
ingest_FEMA6_BLE(
  path_to_ras_dbase,
  HUC8,
  full = FALSE,
  proj_override = NULL,
  vdat_trans = FALSE,
  status_statements = TRUE,
  verbose = FALSE,
  ping_me = NULL,
  quick_check = FALSE,
  quick_hull = FALSE,
  overwrite = FALSE,
  refresh = TRUE
)
}
\arguments{
\item{path_to_ras_dbase}{The path to the folder in which you are building your catalog, Default: NULL}

\item{HUC8}{string to huc8}

\item{full}{flag to download full BLE database, Default: FALSE}

\item{proj_override}{a CRS string to apply should a projection not be found, Default: NULL}

\item{vdat_trans}{a flag to dictate whether or not to apply a vdatum transformation, TRUE to apply, FALSE to skip, Default: FALSE}

\item{status_statements}{flag to determine whether print statements are suppressed, Default: TRUE}

\item{verbose}{flag to determine whether xs statements are suppressed, Default: FALSE}

\item{ping_me}{a string with an email used to send emails after processing. Uses gmailr and requires config, Default: NULL}

\item{quick_check}{a flag to dictate whether or not to perform a quick check to see if a model has already been processed based on the raw name of the file. Useful in reingesting the same directory after an error but not recommended otherwise. TRUE to see if a name matches and skips, FALSE to process all the way though, Default: FALSE}

\item{quick_hull}{a flag to dictate how tightly you want the footprints wrapped, TRUE uses just end points of the linestrings to place the model, FALSE uses the full point database, Default: FALSE}

\item{overwrite}{not currently implemented, Default: FALSE}

\item{refresh}{flag to dictate whether or not to recollate spatial database after ingest process. FALSE to skip, TRUE to regenerate, Default: TRUE}
}
\value{
a RRASSLED catalog of HEC-RAS models
}
\description{
helper to ingest FEMA region 6 BLE data
}
\details{
Hitting https://webapps.usgs.gov/infrm/estBFE/ and https://www.arcgis.com/apps/dashboards/1e98f1e511fc40d3b08790a4251a64ee
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE
 RRASSLER::ingest_FEMA6_BLE(path_to_ras_dbase="G:/data/ras_catalog","12090301",full=FALSE,proj_override = "EPSG:2277",vdat_trans = FALSE,status_statements = TRUE,verbose = FALSE,ping_me = NULL,quick_check = FALSE,quick_hull = FALSE,overwrite = FALSE,refresh = TRUE)
 }
}
}
\seealso{
\code{\link[glue]{glue}}
}
