% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_into_database.R
\name{ingest_into_database}
\alias{ingest_into_database}
\title{ingest_into_database}
\usage{
ingest_into_database(
  path_to_ras_dbase,
  top_of_dir_to_scrape,
  code_to_place_in_source,
  proj_overwrite = NULL,
  vdat_trans = TRUE,
  quiet = FALSE,
  overwrite = FALSE,
  refresh = TRUE
)
}
\arguments{
\item{path_to_ras_dbase}{PARAM_DESCRIPTION}

\item{top_of_dir_to_scrape}{PARAM_DESCRIPTION}

\item{code_to_place_in_source}{PARAM_DESCRIPTION}

\item{proj_overwrite}{PARAM_DESCRIPTION, Default: NULL}

\item{vdat_trans}{PARAM_DESCRIPTION, Default: TRUE}

\item{quiet}{PARAM_DESCRIPTION, Default: FALSE}

\item{overwrite}{PARAM_DESCRIPTION, Default: FALSE}

\item{refresh}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
used to add models from a source folder to a destination catalog
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[as.data.table][data.table::as.data.table], [data.table][data.table::data.table], [rbindlist][data.table::rbindlist], [fwrite][data.table::fwrite]
 [glue][glue::glue]
 [glob2rx][utils::glob2rx]
 [st_crs][sf::st_crs], [st_set_crs][sf::st_set_crs], [st_coordinates][sf::st_coordinates], [st_as_sf][sf::st_as_sf], [st_transform][sf::st_transform], [st_write][sf::st_write]
 [str_sub][stringr::str_sub], [str_detect][stringr::str_detect]
 [unglue_vec][unglue::unglue_vec]
 [sf_linestring][sfheaders::sf_linestring]
 [st_endpoint][lwgeom::st_endpoint], [st_startpoint][lwgeom::st_startpoint]
 [group_by][dplyr::group_by], [distinct][dplyr::distinct]
 [get_nhdplus][nhdplusTools::get_nhdplus]
 [aoi_get][AOI::aoi_get]
 [write_parquet][arrow::write_parquet]
}
