% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_catalog_csv_as_DT.R
\name{load_catalog_csv_as_DT}
\alias{load_catalog_csv_as_DT}
\title{FUNCTION_TITLE}
\usage{
load_catalog_csv_as_DT(path_to_csv, quiet = FALSE)
}
\arguments{
\item{path_to_csv}{PARAM_DESCRIPTION}

\item{quiet}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[fread][data.table::fread]
}
