% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_catalog_csv_as_DT.R
\name{load_catalog_csv_as_DT}
\alias{load_catalog_csv_as_DT}
\title{load_catalog_csv_as_DT}
\usage{
load_catalog_csv_as_DT(path_to_csv, quiet = TRUE)
}
\arguments{
\item{path_to_csv}{path to the model_catalog.csv file under your catalog folder}

\item{quiet}{if TRUE, function will suppress message, Default: TRUE}
}
\value{
returns the model_catalog as a data.table object with enforced column types
}
\description{
Helper to ensure catalog edge cases and incorrect formats are correctly handled (most commonly numeric strings parsed into characters)
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{fread}}
}
