% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_catalog_csv_as_DT.R
\name{load_catalog_csv_as_DT}
\alias{load_catalog_csv_as_DT}
\title{load_catalog_csv_as_DT}
\usage{
load_catalog_csv_as_DT(path_to_csv, is_quiet = TRUE)
}
\arguments{
\item{path_to_csv}{path to the model_catalog.csv file under your catalog folder}

\item{is_quiet}{if TRUE, function will suppress message, Default: TRUE}
}
\value{
returns the model_catalog as a data.table object with enforced column types
}
\description{
Helper to ensure catalog edge cases and incorrect formats are correctly handled (most commonly numeric strings parsed into characters)
}
\details{
a little utility that sets column types on load for some of our sillier edge cases.  the vector applied is `data.table::fread(path_to_csv,colClasses = c("nhdplus_comid" = "character","model_name" = "character","units" = "character","crs" = "character","final_name_key" = "character"))`
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 ras_catalog_dbase = load_catalog_csv_as_DT(file.path(path_to_ras_dbase, "accounting.csv", fsep = .Platform$file.sep),is_quiet = !is_verbose)
 }
}
}
\seealso{
\code{\link[data.table]{fread}}

Other helper: 
\code{\link{marco}()},
\code{\link{print_error_block}()},
\code{\link{print_warning_block}()},
\code{\link{url_exists}()},
\code{\link{util_unzip}()}
}
\concept{helper}
