% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xs_hyfab_comp.R
\name{make_xs_hyfab_comp}
\alias{make_xs_hyfab_comp}
\title{make_xs_hyfab_comp}
\usage{
make_xs_hyfab_comp(path_to_ras_dbase, network_path = NULL)
}
\arguments{
\item{path_to_ras_dbase}{PARAM_DESCRIPTION}

\item{network_path}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
