% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_library.R
\name{map_library}
\alias{map_library}
\title{map_library}
\usage{
map_library(
  path_to_ras_dbase,
  AOI_to_map = NULL,
  name = "model_map",
  plot_lines = FALSE,
  chart_lines = FALSE,
  refresh = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{path_to_ras_dbase}{PARAM_DESCRIPTION}

\item{AOI_to_map}{PARAM_DESCRIPTION, Default: NULL}

\item{name}{PARAM_DESCRIPTION, Default: 'model_map'}

\item{plot_lines}{PARAM_DESCRIPTION, Default: FALSE}

\item{chart_lines}{PARAM_DESCRIPTION, Default: FALSE}

\item{refresh}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
create map of catalog assets
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[st_transform][sf::st_transform], [st_read][sf::st_read], [st_crs][sf::st_crs], [st_make_valid][sf::st_make_valid]
 [read_parquet][arrow::read_parquet]
 [ggplot][ggplot2::ggplot], [aes][ggplot2::aes], [geom_point][ggplot2::geom_point], [theme_light][ggplot2::theme_light], [scale_color_gradientn][ggplot2::scale_color_gradientn], [labs][ggplot2::labs]
 [leaflet][leaflet::leaflet], [leafletOptions][leaflet::leafletOptions], [addProviderTiles][leaflet::addProviderTiles], [addLegend][leaflet::addLegend], [addLayersControl][leaflet::addLayersControl]
 [addFeatures][leafem::addFeatures]
 [popupGraph][leafpop::popupGraph], [popupTable][leafpop::popupTable]
 [glue][glue::glue]
 [mapshot][mapview::mapshot]
}
