% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marco.R
\name{marco}
\alias{marco}
\title{Hello world}
\usage{
marco()
}
\value{
a hello world message (polo!)
}
\description{
A hello world tester
}
\details{
Used to verify that RRASSLER is successfully loaded and test deployment pathways.  Copy-paste-d from Google (https://www.asciiart.eu/art-and-design/sculptures).
}
\examples{
RRASSLER::marco()
}
\seealso{
Other helper: 
\code{\link{load_catalog_csv_as_DT}()},
\code{\link{print_error_block}()},
\code{\link{print_warning_block}()},
\code{\link{process_ras_hdf_to_rasxyz}()},
\code{\link{url_exists}()},
\code{\link{util_unzip}()}
}
\concept{helper}
