% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marco.R, R/move_dbase.R
\name{marco}
\alias{marco}
\alias{move_dbase}
\title{Hello world}
\usage{
marco()

move_dbase(from = NULL, to = NULL)
}
\value{
a hello world message (polo!)

a hello world message (polo!)
}
\description{
A hello world tester

Bulk database mover
}
\details{
No, I did not draw this.  It has taken me longer to document this function than it did to copy-paste it from Google

No, I did not draw this.  It has taken me longer to document this function than it did to copy-paste it from Google
}
\examples{
RRASSLER::marco()
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
Other helper: 
\code{\link{print_error_block}()},
\code{\link{print_warning_block}()},
\code{\link{url_exists}()},
\code{\link{util_unzip}()}

Other helper: 
\code{\link{print_error_block}()},
\code{\link{print_warning_block}()},
\code{\link{url_exists}()},
\code{\link{util_unzip}()}
}
\concept{helper}
