% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_and_plot_old.R
\name{parse_and_plot_old}
\alias{parse_and_plot_old}
\title{parse_and_plot_old}
\usage{
parse_and_plot_old(
  gfile_path = NULL,
  path_to_ras_dbase = NULL,
  excel_row = NULL,
  units_overwrite = NULL,
  proj_overwrite = NULL,
  vdat_trans = FALSE,
  plots_wanted = "g",
  outpath = NULL,
  overwrite = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{gfile_path}{PARAM_DESCRIPTION, Default: NULL}

\item{path_to_ras_dbase}{PARAM_DESCRIPTION, Default: NULL}

\item{excel_row}{PARAM_DESCRIPTION, Default: NULL}

\item{units_overwrite}{PARAM_DESCRIPTION, Default: NULL}

\item{proj_overwrite}{PARAM_DESCRIPTION, Default: NULL}

\item{vdat_trans}{PARAM_DESCRIPTION, Default: TRUE}

\item{plots_wanted}{A string of values representing the autoplots you want from the model.  Valid values include g for g giles h for h files c for combined o for open space to paste ras into parsed values l for a locked aspect ratio, Default: "g"}

\item{outpath}{PARAM_DESCRIPTION, Default: NULL}

\item{overwrite}{PARAM_DESCRIPTION, Default: TRUE}

\item{quiet}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Broken but saving code
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[glue][glue::glue]
 [glob2rx][utils::glob2rx]
 [st_crs][sf::st_crs], [st_set_crs][sf::st_set_crs]
 [sf_linestring][sfheaders::sf_linestring]
 [ggplot][ggplot2::ggplot], [aes][ggplot2::aes], [geom_line][ggplot2::geom_line], [geom_point][ggplot2::geom_point], [theme_light][ggplot2::theme_light], [xlab][ggplot2::xlab], [ylab][ggplot2::ylab], [geom_sf][ggplot2::geom_sf]
 [plot_grid][cowplot::plot_grid], [ggdraw][cowplot::ggdraw], [draw_label][cowplot::draw_label], [save_plot][cowplot::save_plot]
}
