% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_model_to_xyz.R
\name{parse_model_to_xyz}
\alias{parse_model_to_xyz}
\title{parse_model_to_xyz}
\usage{
parse_model_to_xyz(
  geom_path,
  units,
  proj_string,
  in_epoch_override = as.integer(as.POSIXct(Sys.time())),
  out_epoch_override = as.integer(as.POSIXct(Sys.time())),
  vdat_trans = FALSE,
  quiet = FALSE,
  default_g = FALSE,
  try_both = TRUE
)
}
\arguments{
\item{geom_path}{PARAM_DESCRIPTION}

\item{units}{PARAM_DESCRIPTION}

\item{proj_string}{PARAM_DESCRIPTION}

\item{in_epoch_override}{PARAM_DESCRIPTION, Default: as.integer(as.POSIXct(Sys.time()))}

\item{out_epoch_override}{PARAM_DESCRIPTION, Default: as.integer(as.POSIXct(Sys.time()))}

\item{vdat_trans}{PARAM_DESCRIPTION, Default: FALSE}

\item{quiet}{PARAM_DESCRIPTION, Default: FALSE}

\item{default_g}{PARAM_DESCRIPTION, Default: FALSE}

\item{try_both}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_sub}}
 \code{\link[unglue]{unglue}}
}
