% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_error_block.R
\name{print_error_block}
\alias{print_error_block}
\title{print_error_block}
\usage{
print_error_block()
}
\value{
print output
}
\description{
A error block helper
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 RRASSLER::print_error_block()
 }
}
}
