% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_error_block.R, R/print_warning_block.R
\name{print_warning_block}
\alias{print_warning_block}
\title{print_error_block}
\usage{
print_warning_block()

print_warning_block()
}
\value{
print output

print output
}
\description{
A error block helper

A warning block helper
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 RRASSLER::print_error_block()
 }
}
\dontrun{
if(interactive()){
 RRASSLER::print_warning_block()
 }
}
}
