% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_warning_block.R
\name{print_warning_block}
\alias{print_warning_block}
\title{print_warning_block}
\usage{
print_warning_block()
}
\value{
print output
}
\description{
A warning block helper
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 RRASSLER::print_warning_block()
 }
}
}
