% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ras_g_to_xyz.R
\name{process_ras_g_to_xyz}
\alias{process_ras_g_to_xyz}
\title{process_ras_g_to_xyz}
\usage{
process_ras_g_to_xyz(
  geom_path,
  units,
  proj_string,
  in_epoch_override = as.integer(as.POSIXct(Sys.time())),
  out_epoch_override = as.integer(as.POSIXct(Sys.time())),
  vdat = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{geom_path}{path to a file to parse}

\item{units}{units found in the project, "foot" or "meter"}

\item{proj_string}{a projection string to apply}

\item{in_epoch_override}{vdatum parameter input epoch, Default: as.integer(as.POSIXct(Sys.time()))}

\item{out_epoch_override}{vdatum parameter output epoch, Default: as.integer(as.POSIXct(Sys.time()))}

\item{vdat}{a flag to dictate whether or not to apply a vdatum transformation, TRUE to apply, FALSE to skip, Default: FALSE, Default: FALSE}

\item{quiet}{flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE}
}
\value{
a point database and notes about processing
}
\description{
process a ras g file into xyz format
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[read.delim][utils::read.delim]
 [as.data.table][data.table::as.data.table]
 [fill][tidyr::fill]
 [sf_linestring][sfheaders::sf_linestring]
 [st_sf][sf::st_sf], [st_cast][sf::st_cast], [st_crs][sf::st_crs], [st_transform][sf::st_transform], [st_coordinates][sf::st_coordinates], [st_as_sf][sf::st_as_sf], [st_set_crs][sf::st_set_crs], [st_length][sf::st_length]
 [decimal_date][lubridate::decimal_date], [ymd][lubridate::ymd]
 [GET][httr::GET], [http_error][httr::http_error], [content][httr::content]
 [st_linesubstring][lwgeom::st_linesubstring], [st_endpoint][lwgeom::st_endpoint]
 [glue][glue::glue]
}
