% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ras_g_to_xyz.R
\name{process_ras_g_to_xyz}
\alias{process_ras_g_to_xyz}
\title{process_ras_g_to_xyz}
\usage{
process_ras_g_to_xyz(
  geom_path,
  units,
  proj_string,
  in_epoch_override = as.integer(as.POSIXct(Sys.time())),
  out_epoch_override = as.integer(as.POSIXct(Sys.time())),
  vdat = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{geom_path}{path to a file to parse}

\item{units}{units found in the project, "English Units" or "SI Units"}

\item{proj_string}{a projection string to apply}

\item{in_epoch_override}{vdatum parameter input epoch, Default: as.integer(as.POSIXct(Sys.time()))}

\item{out_epoch_override}{vdatum parameter output epoch, Default: as.integer(as.POSIXct(Sys.time()))}

\item{vdat}{a flag to dictate whether or not to apply a vdatum transformation, TRUE to apply, FALSE to skip, Default: FALSE, Default: FALSE}

\item{quiet}{flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE}
}
\value{
a point database and notes about processing
}
\description{
process a ras g file into xyz format
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 # g_path <- "./inst/extdata/sample_ras/FEMA-R6-BLE-sample-dataset/12090301/12090301_models/Model/Alum Creek-Colorado River/ALUM 006/ALUM 006.g01"
 g_path <- fs::path_package("extdata/shapes.fgb", package = "mypkg")
 pts <- process_ras_g_to_xyz(geom_path = g_path,units = "English Units",proj_string = "EPSG:2277",vdat = FALSE,quiet = FALSE)
 }
}
}
\seealso{
\code{\link[utils]{glob2rx}}, \code{\link[utils]{read.table}}
 \code{\link[glue]{glue}}
 \code{\link[data.table]{as.data.table}}
 \code{\link[stringr]{str_split}}, \code{\link[stringr]{str_trim}}, \code{\link[stringr]{str_flatten}}, \code{\link[stringr]{str_sub}}
 \code{\link[sfheaders]{sf_linestring}}
 \code{\link[sf]{sf}}, \code{\link[sf]{st_cast}}, \code{\link[sf]{st_crs}}, \code{\link[sf]{st_read}}, \code{\link[sf]{st_transform}}, \code{\link[sf]{st_coordinates}}, \code{\link[sf]{st_as_sf}}, \code{\link[sf]{geos_measures}}
 \code{\link[tidyr]{fill}}
 \code{\link[lubridate]{decimal_date}}, \code{\link[lubridate]{ymd}}
 \code{\link[httr]{GET}}, \code{\link[httr]{http_error}}, \code{\link[httr]{content}}
 \code{\link[lwgeom]{st_linesubstring}}, \code{\link[lwgeom]{st_startpoint}}
}
