% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_master_files.R
\name{refresh_master_files}
\alias{refresh_master_files}
\title{refresh_master_files}
\usage{
refresh_master_files(path_to_ras_dbase, quiet = TRUE)
}
\arguments{
\item{path_to_ras_dbase}{PARAM_DESCRIPTION}

\item{quiet}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
updates the index and key data objects at the top of the catalog
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
[glob2rx][utils::glob2rx]
 [glue][glue::glue]
 [st_read][sf::st_read], [st_set_crs][sf::st_set_crs], [st_crs][sf::st_crs], [st_write][sf::st_write]
 [read_parquet][arrow::read_parquet], [write_parquet][arrow::write_parquet]
 [as.data.table][data.table::as.data.table], [rbindlist][data.table::rbindlist]
 [sf_linestring][sfheaders::sf_linestring]
}
