% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_master_files.R
\name{refresh_master_files}
\alias{refresh_master_files}
\title{refresh_master_files}
\usage{
refresh_master_files(path_to_ras_dbase, point_merge = TRUE, verbose = TRUE)
}
\arguments{
\item{path_to_ras_dbase}{The path to the folder in which you are building your catalog, Default: NULL}

\item{point_merge}{Dev flag to shave point processing, Default: TRUE}

\item{verbose}{flag to determine whether print statements are shown - TRUE to show messages - FALSE to skip non-critical ones, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
remerge individual files into spatial model key
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE
 RRASSLER::refresh_master_files(path_to_ras_dbase = "G:/data/ras_catalog",verbose=TRUE)
 }
}
}
\seealso{
\code{\link[utils]{glob2rx}}
 \code{\link[glue]{glue}}
 \code{\link[sf]{st_read}}, \code{\link[sf]{st_crs}}, \code{\link[sf]{st_write}}
 \code{\link[arrow]{read_parquet}}, \code{\link[arrow]{write_parquet}}
 \code{\link[data.table]{as.data.table}}, \code{\link[data.table]{rbindlist}}
 \code{\link[sfheaders]{sf_linestring}}
}
