% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_master_files.R
\name{refresh_master_files}
\alias{refresh_master_files}
\title{refresh_master_files}
\usage{
refresh_master_files(path_to_ras_dbase, quiet = TRUE)
}
\arguments{
\item{path_to_ras_dbase}{The path to the folder in which you are building your catalog, Default: NULL}

\item{quiet}{flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE, Default: TRUE}
}
\value{
a new set of spatial index files
}
\description{
remerge individual files into spatial model key
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 RRASSLER::refresh_master_files(path_to_ras_dbase = "G:/data/ras_catalog",quiet=FALSE)
 }
}
}
\seealso{
\code{\link[utils]{glob2rx}}
 \code{\link[glue]{glue}}
 \code{\link[sf]{st_read}}, \code{\link[sf]{st_crs}}, \code{\link[sf]{st_write}}
 \code{\link[arrow]{read_parquet}}, \code{\link[arrow]{write_parquet}}
 \code{\link[data.table]{as.data.table}}, \code{\link[data.table]{rbindlist}}
 \code{\link[sfheaders]{sf_linestring}}
}
