% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRASSLER-package.R
\docType{package}
\name{RRASSLER-package}
\alias{RRASSLER}
\alias{RRASSLER-package}
\title{RRASSLER: R based HEC-RAS model accounting wrestler}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A package for making HEC-RAS data more FAIR by cataloging models, cross sections, and points for accounting, interoperability, and reuse in RAS and other applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NOAA-OWP/RRASSLER}
  \item Report bugs at \url{https://github.com/NOAA-OWP/RRASSLER/issues/new}
}

}
\author{
\strong{Maintainer}: Jim Coll \email{james.coll@noaa.gov} (\href{https://orcid.org/0000-0002-6994-6668}{ORCID})

Authors:
\itemize{
  \item Mike Johnson \email{mike.johnson@noaa.gov} (\href{https://orcid.org/0000-0002-5288-8350}{ORCID})
}

Other contributors:
\itemize{
  \item Laura Keys (\href{https://orcid.org/0000-0003-1470-5215}{ORCID}) [contributor]
  \item NOAA [funder]
  \item Lynker [funder]
}

}
\keyword{internal}
