% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_models_from_table.R
\name{update_models_from_table}
\alias{update_models_from_table}
\title{update_models_from_table}
\usage{
update_models_from_table(path_to_ras_dbase, refresh = FALSE, quiet = TRUE)
}
\arguments{
\item{path_to_ras_dbase}{The path to the top level folder of the ras_catalog structure, Default: TRUE}

\item{refresh}{flag to dictate whether or not to recollate spatial database prior to mapping.  FALSE to skip, TRUE to regenerate, Default: TRUE}

\item{quiet}{flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE}
}
\value{
a reRRASSLED catalog
}
\description{
try and reprocess the _unprocessed models
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringi]{stri_isempty}}
 \code{\link[glue]{glue}}
 \code{\link[stringr]{str_detect}}
 \code{\link[data.table]{data.table-package}}, \code{\link[data.table]{rbindlist}}, \code{\link[data.table]{fwrite}}
 \code{\link[sfheaders]{sf_linestring}}
 \code{\link[sf]{st_crs}}, \code{\link[sf]{st_coordinates}}, \code{\link[sf]{st_as_sf}}, \code{\link[sf]{st_transform}}, \code{\link[sf]{geos_unary}}, \code{\link[sf]{geos_combine}}, \code{\link[sf]{st_write}}
 \code{\link[lwgeom]{st_startpoint}}
 \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{distinct}}
 \code{\link[nhdplusTools]{get_nhdplus}}
 \code{\link[AOI]{aoi_get}}
 \code{\link[arrow]{write_parquet}}
}
