% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_g_to_geom_pts.R
\name{util_g_to_geom_pts}
\alias{util_g_to_geom_pts}
\title{util_g_to_geom_pts}
\usage{
util_g_to_geom_pts(geom_path, units, proj_string, quiet = FALSE)
}
\arguments{
\item{geom_path}{path to a file to parse}

\item{units}{units found in the project, "English Units" or "SI Units"}

\item{proj_string}{a projection string to apply}

\item{quiet}{flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE}
}
\value{
sf points
}
\description{
A wrapper of a few different utilities across rrassler to read files in directly as points
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[utils]{glob2rx}}, \code{\link[utils]{read.table}}
 \code{\link[glue]{glue}}
 \code{\link[data.table]{as.data.table}}
 \code{\link[stringr]{str_split}}, \code{\link[stringr]{str_trim}}, \code{\link[stringr]{str_flatten}}
 \code{\link[sfheaders]{sf_linestring}}
 \code{\link[sf]{sf}}, \code{\link[sf]{st_cast}}
 \code{\link[tidyr]{fill}}
}
