% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_unzip.R
\name{util_unzip}
\alias{util_unzip}
\title{unzip nested folders}
\usage{
util_unzip(zippath, is_quiet = FALSE)
}
\arguments{
\item{zippath}{path to zip}

\item{is_quiet}{flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE}
}
\value{
unzipped dir
}
\description{
unzips nested zip files
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 RRASSLER::util_unzip(file.path(database_path,"_temp","BLE",HUCID,glue::glue("{HUCID}_models.zip"),fsep = .Platform$file.sep),is_quiet = is_quiet)
 }
}
}
\seealso{
\code{\link[stringr]{str_sub}}
 \code{\link[utils]{unzip}}
 \code{\link[glue]{glue}}
 \code{\link[dplyr]{setops}}

Other helper: 
\code{\link{load_catalog_csv_as_DT}()},
\code{\link{marco}()},
\code{\link{print_error_block}()},
\code{\link{print_warning_block}()},
\code{\link{url_exists}()}
}
\concept{helper}
