% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_url.R
\name{url_exists}
\alias{url_exists}
\title{URL tester}
\usage{
url_exists(x, non_2xx_return_value = FALSE, is_quiet = FALSE)
}
\arguments{
\item{x}{the URL to test}

\item{non_2xx_return_value}{PARAM_DESCRIPTION, Default: FALSE}

\item{is_quiet}{flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE}
}
\value{
returns TRUE if the url exists
}
\description{
A function to test if I get a response from a URL
}
\details{
gratefully pilfered from https://stackoverflow.com/questions/52911812/check-if-url-exists-in-r
}
\examples{
#EXAMPLE1
some_urls <-
c(
  "https://jimcoll.github.io/a/bad/programmer/",    # Should 404
  "https://github.com/NOAA-OWP/inundation-mapping", # Should exist
  "https://nooneshouldeverhavethisurl.gov/badtest"  # Does not exist in any form
)
data.frame(
  exists = sapply(some_urls, url_exists, USE.NAMES = FALSE),
  some_urls,
  stringsAsFactors = FALSE
) \%>\% tibble::as_tibble() \%>\% print()
}
\seealso{
\code{\link[httr]{HEAD}}, \code{\link[httr]{GET}}, \code{\link[httr]{status_code}}

Other helper: 
\code{\link{load_catalog_csv_as_DT}()},
\code{\link{marco}()},
\code{\link{print_error_block}()},
\code{\link{print_warning_block}()},
\code{\link{process_ras_hdf_to_rasxyz}()},
\code{\link{util_unzip}()}
}
\concept{helper}
