% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyz_to_ls.R
\name{xyz_to_ls}
\alias{xyz_to_ls}
\title{xyz_to_ls}
\usage{
xyz_to_ls(dat = NULL, path_to_ras_dbase = NULL, is_quiet = FALSE)
}
\arguments{
\item{dat}{The dataframe of points, Default: NULL}

\item{path_to_ras_dbase}{The path to the folder in which you are building your catalog, is also location agnostic (disk or cloud), Default: NULL}

\item{is_quiet}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
linestrings
}
\description{
helper to transform points into linestrings
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[arrow]{read_parquet}}
 \code{\link[sfheaders]{sf_linestring}}
 \code{\link[sf]{st_crs}}
}
